support github CI workflow (#96)
* support github cicd workflow


* fix code checker workflow


* fix code format lint error


* disable extra-rules code format


* fix golang lint error


* fix unittest datarace err


* fix ut datarace


* fix discovery perf


* Split the integration test into multiple subset tests and execute them concurrently


* change sub ITs job names


* Split the e2e test set


* improve workflow names


* remove writer etcd test


* improve workload log output


* fix create log already exists test assert


* The e2e test workflow filters the corresponding test subsets


* fix wait minicluster stop loop


* improve test log output


* improve cicd workflow log output


* add ci-passed label & rerun trigger


* fix comment not trigger rerun


---------

+++ woodpecker/quorum/discovery.go
@@ -174,7 +174,7 @@ func (d *quorumDiscovery) buildCrossRegionFilters(ctx context.Context) error {
 	// Create one filter that will be used for each region
 	d.filters = []*proto.NodeFilter{
 		{
-			Limit: int32(d.es), // Will be adjusted per region during selection
+			Limit: d.es, // Will be adjusted per region during selection
 		},
 	}
 
@@ -219,7 +219,6 @@ func (d *quorumDiscovery) SelectQuorum(ctx context.Context) (*proto.QuorumInfo, 
 		}
 		return err
 	}, retry.AttemptAlways(), retry.Sleep(200*time.Millisecond), retry.MaxSleepTime(2*time.Second))
-
 	if err != nil {
 		return nil, werr.ErrServiceSelectQuorumFailed.WithCauseErr(err)
 	}
